fix prompt reduction
