+++ src/main/resources/messages/TestGenie.properties
@@ -8,8 +8,6 @@ compilationTestsChecking=Compilation tests checking
 runningJacoco=Running jacoco
 testCasesSaving=Test cases saving
 calculatingCoverage=Calculating coverage
-polyDepthReducing=Reducing the polygons depth to
-inputParamsDepthReducing=Reducing the input params depth to
 !Build error titles
 buildErrorTitle=Build error
 evosuiteErrorTitle=EvoSuite error
@@ -53,6 +51,6 @@ serverProblems=Grazie Platform is not responding
 tooLongPrompt=The generated prompt is too long! Please, generate tests for single methods.
 requestError=An error occurred during interaction with the Grazie Platform. Code:
 savingTestFileIssue=Plugin cannot save the test file
-promptReduction=The generated prompt is too long, plugin will reduce Large Language Model parameters
+promptReduction=The generated prompt is too long, plugin reduced Large Language Model parameters.
 !Build error messages
 commonBuildErrorMessage=Please make sure that IntelliJ can build your project without any issues or provide the correct build command in the settings
